Fix warning on deleteLater on nullptr in QSampleCache
It is possible to have a loading thread running, but with no network
manager.
In that case, calling deleteLater() on m_networkAccessManager results in
a warning: postEvent: Unexpected null receiver.

Fix this issue by checking whether the network manager exists before
calling deleteLater() on it.

Task-number: QTBUG-66910
(cherry picked from commit 448c992972d4078af1eee7ad9ccc8d86c0f2e78d)

+++ src/multimedia/audio/qsamplecache_p.cpp
@@ -141,8 +141,10 @@ void QSampleCache::loadingRelease()
     m_loadingRefCount--;
     if (m_loadingRefCount == 0) {
         if (m_loadingThread.isRunning()) {
-            m_networkAccessManager->deleteLater();
-            m_networkAccessManager = nullptr;
+            if (m_networkAccessManager) {
+                m_networkAccessManager->deleteLater();
+                m_networkAccessManager = nullptr;
+            }
             m_loadingThread.exit();
         }
     }
